fix: Fix shapemap default fail color to be RED
- Move NoMatchReason from validator_error.rs into shex_validation::no_match_reason
- Clarify ConditionFailed/rbe_error wording ("No candidates to match. Missing: ...")
- Fix ShapemapConfig::default_fail_color(), which returned Green instead of Red
