+++ shex_validation/src/validator_error.rs
@@ -1,9 +1,9 @@
 use crate::PartitionsDisplay;
 use crate::Reasons;
 use crate::ValidatorErrors;
+use crate::no_match_reason::NoMatchReason;
 use prefixmap::PrefixMap;
 use prefixmap::error::PrefixMapError;
-use rbe::Cardinality;
 use rbe::RbeError;
 use rudof_iri::IriS;
 use rudof_rdf::rdf_core::term::Object;
@@ -18,76 +18,6 @@ use shex_ast::ir::shape_expr::ShapeExpr;
 use shex_ast::{Node, Pred, ShapeExprLabel, ShapeLabelIdx, ast::cond_kind::CondKind, ir::shape_label::ShapeLabel};
 use termtree::Tree;
 use thiserror::Error;
-
-/// Why a single candidate assignment of neighbors to a triple expression was
-/// rejected, attached as children of a [`ValidatorError::NoMatchesFound`]
-/// error so the user can see, per candidate, why it didn't work.
-#[derive(Debug, Clone)]
-pub enum NoMatchReason {
-    /// `value` didn't satisfy `predicate`'s own condition (e.g. a node
-    /// constraint or shape reference).
-    ConditionFailed {
-        candidate: Vec<(Pred, Node)>,
-        predicate: Pred,
-        value: Node,
-        error: RbeError<Pred, Node, ShapeLabelIdx, SemanticActionContext, CondKind>,
-    },
-    /// `predicate` needed to occur `expected` times but occurred `current`
-    /// times among the candidate's neighbors.
-    CardinalityFailed {
-        candidate: Vec<(Pred, Node)>,
-        predicate: Pred,
-        expected: Cardinality,
-        current: usize,
-    },
-    /// A candidate was rejected for a reason that couldn't be attributed to
-    /// a single predicate's cardinality (e.g. `Or`-branch interactions).
-    Other {
-        candidate: Vec<(Pred, Node)>,
-        detail: String,
-    },
-}
-
-impl NoMatchReason {
-    fn show_qualified(&self, nodes_prefixmap: &PrefixMap) -> String {
-        let show_pred = |p: &Pred| nodes_prefixmap.qualify(p.iri());
-        let show_node = |n: &Node| n.show_qualified(nodes_prefixmap);
-        let show_candidate = |candidate: &[(Pred, Node)]| {
-            candidate
-                .iter()
-                .map(|(p, v)| format!("{} {}", show_pred(p), v.show_qualified(nodes_prefixmap)))
-                .collect::<Vec<_>>()
-                .join(", ")
-        };
-        match self {
-            NoMatchReason::ConditionFailed {
-                candidate,
-                predicate,
-                value,
-                error,
-            } => format!(
-                "Candidate [{}] rejected: {} {}: {}",
-                show_candidate(candidate),
-                show_pred(predicate),
-                value.show_qualified(nodes_prefixmap),
-                error.show_qualified(&show_pred, &show_node),
-            ),
-            NoMatchReason::CardinalityFailed {
-                candidate,
-                predicate,
-                expected,
-                current,
-            } => format!(
-                "Candidate [{}] rejected: predicate {} required cardinality {expected:?} but got {current}",
-                show_candidate(candidate),
-                show_pred(predicate),
-            ),
-            NoMatchReason::Other { candidate, detail } => {
-                format!("Candidate [{}] rejected: {detail}", show_candidate(candidate))
-            },
-        }
-    }
-}
 
 #[derive(Error, Debug, Clone)]
 pub enum ValidatorError {
@@ -632,8 +562,8 @@ impl ValidatorError {
         schema: &SchemaIR,
         width: usize,
     ) -> Result<String, PrefixMapError> {
-        // A shape made of a single triple constraint (the overwhelmingly common
-        // case) rejected on its one and only candidate reduces to a single
+        // A shape made of a single triple constraint rejected on its one
+        // and only candidate reduces to a single
         // `ConditionFailed` reason. Render that flat, naming the node and
         // property involved, instead of the generic
         // "Shape ... failed ...: no candidates matched" + one-line tree.
